+++ CRM/Smartdebit/Base.php
@@ -125,7 +125,7 @@ WHERE ddi_reference = '{$params['ddi_reference']}'
     $ddDetails['account_holder'] = CRM_Utils_Array::value('account_holder', $params);
     $ddDetails['bank_account_number'] = CRM_Utils_Array::value('bank_account_number', $params);
     $ddDetails['bank_identification_number'] = CRM_Utils_Array::value('bank_identification_number', $params);
-    $ddDetails['bank_name'] = CRM_Utils_Array::value('bank_name', $params);
+    $ddDetails['bank_name'] = CRM_Utils_Array::value('bank_name', $params, $ddDetails['bank_name']);
 
     $ddDetails['sun'] = CRM_Smartdebit_Base::getSUN();